+++ sovtoken/setup.py
@@ -38,7 +38,7 @@ setup(
              '*.css', '*.ico', '*.png', 'LICENSE', 'LEGAL', 'sovtoken']},
     include_package_data=True,
 
-    install_requires=['indy-node>=1.7.0.dev893'],
+    install_requires=['indy-node>=1.7.0.dev894'],
 
     setup_requires=['pytest-runner'],
     extras_require={
