+++ sovtokenfees/sovtokenfees/test/catchup/test_xfer_nym_during_catchup.py
@@ -4,7 +4,7 @@ from indy_common.constants import NYM
 
 from sovtoken.constants import XFER_PUBLIC
 
-from sovtokenfees.test.view_change.helper import scenario_txns_during_view_change
+from sovtokenfees.test.catchup.helper import scenario_txns_during_catchup
 
 
 @pytest.fixture(
@@ -20,11 +20,11 @@ def fees(request):
     return request.param
 
 
-def test_xfer_nym_during_view_change(
-        looper,
+def test_revert_xfer_and_other_txn_before_catchup(
+        looper, tconf, tdir, allPluginsPath,
+        do_post_node_creation,
         nodeSetWithIntegratedTokenPlugin,
         fees_set,
-        curr_utxo,
         send_and_check_transfer_curr_utxo,
         send_and_check_nym_with_fees_curr_utxo,
 ):
@@ -32,4 +32,8 @@ def test_xfer_nym_during_view_change(
         send_and_check_transfer_curr_utxo()
         send_and_check_nym_with_fees_curr_utxo()
 
-    scenario_txns_during_view_change(looper, nodeSetWithIntegratedTokenPlugin, send_txns, curr_utxo)
+    scenario_txns_during_catchup(
+        looper, tconf, tdir, allPluginsPath, do_post_node_creation,
+        nodeSetWithIntegratedTokenPlugin,
+        send_txns
+    )
