+++ sovtokenfees/sovtokenfees/test/view_change/test_nym_during_view_change.py
@@ -1,10 +1,19 @@
+import pytest
+
 from sovtokenfees.test.view_change.helper import scenario_txns_during_view_change
 
 from indy_common.constants import NYM
 
 
-def fees():
-    return {NYM: 0}  # no fees
+@pytest.fixture(
+    scope='module',
+    params=[
+        {NYM: 0},  # no fees
+        {NYM: 4},  # with fees
+    ], ids=lambda x: 'fees' if x[NYM] else 'nofees'
+)
+def fees(request):
+    return request.param
 
 
 def test_nym_during_view_change(
