+++ sovtokenfees/sovtokenfees/test/view_change/test_xfer_during_view_change.py
@@ -1,10 +1,19 @@
+import pytest
+
 from sovtokenfees.test.view_change.helper import scenario_txns_during_view_change
 
 from sovtoken.constants import XFER_PUBLIC
 
 
-def fees():
-    return {XFER_PUBLIC: 0}  # no fees
+@pytest.fixture(
+    scope='module',
+    params=[
+        {XFER_PUBLIC: 0},  # no fees
+        {XFER_PUBLIC: 4},  # with fees
+    ], ids=lambda x: 'fees' if x[XFER_PUBLIC] else 'nofees'
+)
+def fees(request):
+    return request.param
 
 
 def test_xfer_during_view_change(
